+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package are documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2025-07-31
+### Added
+ - HDR support, input textures are now detected as HDR, and output texture can now be set to be HDR
+ - Force include alpha toggle, for when you want to include a solid 0 or 1 alpha channel by leaving the slot empty
+
 ## [0.4.3] - 2025-07-31
 ### Added
  - Multi editing support